src/main/java/com/stackroute/exercise1/vowel.java
package com.stackroute.exercise1;


public class vowel
{
    public String vowelConsonantCheck(String word)
    {
        String var="";
        for(int i=0;i<word.length();++i)
        {
            char ch= word.charAt(i);
            if(ch=='a' || ch=='e' || ch=='i' || ch=='o' || ch=='u')
            {
                var+="Vowel";
            }
            else if (ch>='a' && ch<='z')
            {
                var+="Consonant";
            }
            else
            {
                System.out.print("Please enter a valid letter");
            }
        }
        return var;
    }
}
